Prepare a working example
Add the instana-config.json as well. Get and patch the NGINX config
template and provide it from the init container as well. The idea
is to configure the tracer via environment variables. So those need
to be whitelisted in the NGINX config template.

Provide the instructions for the Docker image upload to a registry
and to use the image from there.