Enhance the build with a container upload draft
Move all build files to "build/". Add a script to get the current
SaaS release updating the release "build/Jenkinsfile" and reminding
the caller to update the build offset in there as well. Calculate a
minor number by build number - build offset.
Pass a version 1.${releaseNumber}.${minorNumber} to the new upload
script. Just print the first docker command that should be executed
in the upload script for now.